+++ booklog/src/App.tsx
@@ -1,11 +1,12 @@
 import { Header } from "./components/Header/Header";
 import { Body } from "./components/Body/Body";
+import { BooksContextProvider } from "./context";
 
 export const App = (): JSX.Element => {
   return (
-    <>
+    <BooksContextProvider>
       <Header />
       <Body />
-    </>
+    </BooksContextProvider>
   );
 };
