+++ booklog/src/components/Body/Body.tsx
@@ -1,7 +1,8 @@
 import { useContext } from "react";
-import { BooksContext } from "../../context/BooksProvider";
-import { MyBooksProvider } from "../../context/MyBooksProvider";
-import { BooksProvider } from "../../context/BooksProvider";
+
+//context
+import { BooksContext } from "../../context";
+
 //component
 import { Books } from "./Books/Books";
 import { SearchBar } from "./SearchBar/SearchBar";
@@ -13,16 +14,12 @@ import styles from "../../App.module.css";
 export const Body = (): JSX.Element => {
   const { bookItems } = useContext(BooksContext);
   return (
-    <MyBooksProvider>
-      <div className={styles.conWrap}>
-        <SideBar />
-        <BooksProvider>
-          <main className={styles.boxMain}>
-            <SearchBar />
-            {bookItems.length !== 0 && <Books />}
-          </main>
-        </BooksProvider>
-      </div>
-    </MyBooksProvider>
+    <div className={styles.conWrap}>
+      <SideBar />
+      <main className={styles.boxMain}>
+        <SearchBar />
+        {bookItems.length !== 0 && <Books />}
+      </main>
+    </div>
   );
 };
