+++ booklog/src/components/Body/SearchBar/SearchBar.tsx
@@ -1,10 +1,9 @@
 import { useRef, useContext } from "react";
-import { BooksContext } from "../../../context/BooksProvider";
+import { BooksContext } from "../../../context";
 import styles from "./SearchBar.module.css";
 
 export const SearchBar = (): JSX.Element => {
   const { fetchBooksApi } = useContext(BooksContext);
-
   const inputRef = useRef<HTMLInputElement>(null);
   const handleSearch = (e: React.MouseEvent<HTMLButtonElement>) => {
     e.preventDefault();
@@ -14,7 +13,6 @@ export const SearchBar = (): JSX.Element => {
       alert("キーワード入力してください！");
     }
   };
-
   return (
     <form className={styles.boxSearchBar} action="">
       <input
