+++ booklog/src/types/index.ts
@@ -21,13 +21,10 @@ export type BookItem = {
 };
 
 export type MyBooksContextType = {
-  myBooks: BookItem[];
-  addMyBooks: (selectedBook: BookItem) => void;
-  deleteMyBooks: (id: string) => void;
-};
-
-export type BooksContextType = {
   total: number;
   bookItems: BooksResult["items"];
+  myBooks: BookItem[];
   fetchBooksApi: (query: string) => void;
+  addMyBooks: (selectedBook: BookItem) => void;
+  deleteMyBooks: (id: string) => void;
 };
